fix: run only on global installation

+++ src/services/analytics/postinstall.ts
@@ -1,18 +1,22 @@
 import * as fs from 'fs'
+import { env } from 'process'
 import { analyticsService } from '.'
 import { EventDTO } from './analytics.api'
 
-const check = fs.readFileSync('./file.txt', 'utf-8')
-const analyticsData: EventDTO = {
-  name: 'CLI_ANALYTICS_ENABLED',
-  category: 'APPLICATION',
-  component: 'Cli',
-  uuid: '',
-  metadata: {},
+const isGlobal = env.npm_config_global === 'true'
+if (isGlobal) {
+  const check = fs.readFileSync('./file.txt', 'utf-8')
+  const analyticsData: EventDTO = {
+    name: 'CLI_ANALYTICS_ENABLED',
+    category: 'APPLICATION',
+    component: 'Cli',
+    uuid: '',
+    metadata: {},
+  }
+  if (check === 'installation') {
+    analyticsService.eventsControllerSend(analyticsData)
+  }
+  fs.unlink('./file.txt', (err) => {
+    if (err) throw err
+  })
 }
-if (check === 'installation') {
-  analyticsService.eventsControllerSend(analyticsData)
-}
-fs.unlink('./file.txt', (err) => {
-  if (err) throw err
-})
